modify hierarchy for GCP
